+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/TileGenerator.java
@@ -12,11 +12,15 @@ import org.locationtech.proj4j.units.Units;
 
 import java.io.IOException;
 import java.sql.SQLException;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 
 import mil.nga.geopackage.BoundingBox;
 import mil.nga.geopackage.GeoPackage;
@@ -67,14 +71,9 @@ public abstract class TileGenerator {
     private final String tableName;
 
     /**
-     * Min zoom level
+     * Zoom levels
      */
-    private final int minZoom;
-
-    /**
-     * Max zoom level
-     */
-    private final int maxZoom;
+    private SortedSet<Integer> zoomLevels = new TreeSet<>();
 
     /**
      * Tiles projection
@@ -159,6 +158,43 @@ public abstract class TileGenerator {
      * @param context     app context
      * @param geoPackage  GeoPackage
      * @param tableName   table name
+     * @param boundingBox tiles bounding box
+     * @param projection  tiles projection
+     * @since 6.1.4
+     */
+    public TileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                         BoundingBox boundingBox, Projection projection) {
+        this.context = context;
+        geoPackage.verifyWritable();
+        this.geoPackage = geoPackage;
+        this.tableName = tableName;
+        this.boundingBox = boundingBox;
+        this.projection = projection;
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context     app context
+     * @param geoPackage  GeoPackage
+     * @param tableName   table name
+     * @param zoomLevel   zoom level
+     * @param boundingBox tiles bounding box
+     * @param projection  tiles projection
+     * @since 6.1.4
+     */
+    public TileGenerator(Context context, GeoPackage geoPackage, String tableName, int zoomLevel,
+                         BoundingBox boundingBox, Projection projection) {
+        this(context, geoPackage, tableName, boundingBox, projection);
+        addZoomLevel(zoomLevel);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context     app context
+     * @param geoPackage  GeoPackage
+     * @param tableName   table name
      * @param minZoom     min zoom
      * @param maxZoom     max zoom
      * @param boundingBox tiles bounding box
@@ -167,15 +203,43 @@ public abstract class TileGenerator {
      */
     public TileGenerator(Context context, GeoPackage geoPackage,
                          String tableName, int minZoom, int maxZoom, BoundingBox boundingBox, Projection projection) {
-        this.context = context;
-        geoPackage.verifyWritable();
-        this.geoPackage = geoPackage;
-        this.tableName = tableName;
+        this(context, geoPackage, tableName, boundingBox, projection);
+        addZoomLevels(minZoom, maxZoom);
+    }
 
-        this.minZoom = minZoom;
-        this.maxZoom = maxZoom;
-        this.boundingBox = boundingBox;
-        this.projection = projection;
+    /**
+     * Constructor
+     *
+     * @param context     app context
+     * @param geoPackage  GeoPackage
+     * @param tableName   table name
+     * @param zoomLevels  zoom levels
+     * @param boundingBox tiles bounding box
+     * @param projection  tiles projection
+     * @since 6.1.4
+     */
+    public TileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                         Collection<Integer> zoomLevels, BoundingBox boundingBox,
+                         Projection projection) {
+        this(context, geoPackage, tableName, boundingBox, projection);
+        addZoomLevels(zoomLevels);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context     app context
+     * @param geoPackage  GeoPackage
+     * @param tableName   table name
+     * @param zoomLevels  zoom levels
+     * @param boundingBox tiles bounding box
+     * @param projection  tiles projection
+     * @since 6.1.4
+     */
+    public TileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                         int[] zoomLevels, BoundingBox boundingBox, Projection projection) {
+        this(context, geoPackage, tableName, boundingBox, projection);
+        addZoomLevels(zoomLevels);
     }
 
     /**
@@ -205,7 +269,8 @@ public abstract class TileGenerator {
      * @since 1.2.5
      */
     public int getMinZoom() {
-        return minZoom;
+        validateZoomLevels();
+        return zoomLevels.first();
     }
 
     /**
@@ -215,7 +280,71 @@ public abstract class TileGenerator {
      * @since 1.2.5
      */
     public int getMaxZoom() {
-        return maxZoom;
+        validateZoomLevels();
+        return zoomLevels.last();
+    }
+
+    /**
+     * Get the zoom levels (read only)
+     *
+     * @return zoom levels
+     * @since 6.1.4
+     */
+    public SortedSet<Integer> getZoomLevels() {
+        return Collections.unmodifiableSortedSet(zoomLevels);
+    }
+
+    /**
+     * Add a zoom level
+     *
+     * @param zoomLevel zoom level
+     * @return true if zoom level added
+     * @since 6.1.4
+     */
+    public boolean addZoomLevel(int zoomLevel) {
+        return zoomLevels.add(zoomLevel);
+    }
+
+    /**
+     * Add a range of zoom levels
+     *
+     * @param minZoom min zoom level
+     * @param maxZoom max zoom level
+     * @return true if at least one zoom level added
+     * @since 6.1.4
+     */
+    public boolean addZoomLevels(int minZoom, int maxZoom) {
+        boolean added = false;
+        for (int zoomLevel = minZoom; zoomLevel <= maxZoom; zoomLevel++) {
+            added = addZoomLevel(zoomLevel) || added;
+        }
+        return added;
+    }
+
+    /**
+     * Add zoom levels
+     *
+     * @param zoomLevels zoom levels
+     * @return true if at least one zoom level added
+     * @since 6.1.4
+     */
+    public boolean addZoomLevels(Collection<Integer> zoomLevels) {
+        return this.zoomLevels.addAll(zoomLevels);
+    }
+
+    /**
+     * Add zoom levels
+     *
+     * @param zoomLevels zoom levels
+     * @return true if at least one zoom level added
+     * @since 6.1.4
+     */
+    public boolean addZoomLevels(int[] zoomLevels) {
+        boolean added = false;
+        for (int zoomLevel : zoomLevels) {
+            added = addZoomLevel(zoomLevel) || added;
+        }
+        return added;
     }
 
     /**
@@ -379,6 +508,9 @@ public abstract class TileGenerator {
      */
     public int getTileCount() {
         if (tileCount == null) {
+
+            validateZoomLevels();
+
             long count = 0;
 
             boolean degrees = projection.isUnit(Units.DEGREES);
@@ -387,7 +519,7 @@ public abstract class TileGenerator {
                 transformToWebMercator = projection.getTransformation(ProjectionConstants.EPSG_WEB_MERCATOR);
             }
 
-            for (int zoom = minZoom; zoom <= maxZoom; zoom++) {
+            for (int zoom : zoomLevels) {
 
                 BoundingBox expandedBoundingBox = getBoundingBox(zoom);
 
@@ -418,6 +550,8 @@ public abstract class TileGenerator {
      */
     public int generateTiles() throws SQLException, IOException {
 
+        validateZoomLevels();
+
         int totalCount = getTileCount();
 
         // Set the max progress count
@@ -429,6 +563,8 @@ public abstract class TileGenerator {
         boolean update = false;
 
         // Adjust the tile matrix set and bounds
+        int minZoom = getMinZoom();
+        int maxZoom = getMaxZoom();
         BoundingBox minZoomBoundingBox = tileBounds.get(minZoom);
         adjustBounds(minZoomBoundingBox, minZoom);
 
@@ -475,27 +611,31 @@ public abstract class TileGenerator {
             for (int zoom = minZoom; zoom <= maxZoom
                     && (progress == null || progress.isActive()); zoom++) {
 
-                TileGrid localTileGrid = null;
+                if (zoomLevels.contains(zoom)) {
 
-                // Determine the matrix width and height for XYZ format
-                if (xyzTiles) {
-                    matrixWidth = TileBoundingBoxUtils.tilesPerSide(zoom);
-                    matrixHeight = matrixWidth;
-                }
-                // Get the local tile grid for GeoPackage format of where the
-                // tiles belong
-                else {
-                    BoundingBox zoomBoundingBox = tileBounds.get(zoom);
-                    localTileGrid = TileBoundingBoxUtils.getTileGrid(
-                            tileGridBoundingBox, matrixWidth, matrixHeight,
-                            zoomBoundingBox);
-                }
+                    TileGrid localTileGrid = null;
 
-                // Generate the tiles for the zoom level
-                TileGrid tileGrid = tileGrids.get(zoom);
-                count += generateTiles(tileMatrixDao, tileDao, contents, zoom,
-                        tileGrid, localTileGrid, matrixWidth, matrixHeight,
-                        update);
+                    // Determine the matrix width and height for XYZ format
+                    if (xyzTiles) {
+                        matrixWidth = TileBoundingBoxUtils.tilesPerSide(zoom);
+                        matrixHeight = matrixWidth;
+                    }
+                    // Get the local tile grid for GeoPackage format of where the
+                    // tiles belong
+                    else {
+                        BoundingBox zoomBoundingBox = tileBounds.get(zoom);
+                        localTileGrid = TileBoundingBoxUtils.getTileGrid(
+                                tileGridBoundingBox, matrixWidth, matrixHeight,
+                                zoomBoundingBox);
+                    }
+
+                    // Generate the tiles for the zoom level
+                    TileGrid tileGrid = tileGrids.get(zoom);
+                    count += generateTiles(tileMatrixDao, tileDao, contents, zoom,
+                            tileGrid, localTileGrid, matrixWidth, matrixHeight,
+                            update);
+
+                }
 
                 if (!xyzTiles) {
                     // Double the matrix width and height for the next level
@@ -527,6 +667,16 @@ public abstract class TileGenerator {
         }
 
         return count;
+    }
+
+    /**
+     * Validate that at least one zoom level was specified
+     */
+    private void validateZoomLevels() {
+        if (zoomLevels.isEmpty()) {
+            throw new GeoPackageException(
+                    "At least one zoom level must be specified");
+        }
     }
 
     /**
@@ -655,6 +805,7 @@ public abstract class TileGenerator {
             // Adjust the bounds to include the request and existing bounds
             ProjectionTransform transformProjectionToTileMatrixSet = projection.getTransformation(tileMatrixProjection);
             boolean sameProjection = transformProjectionToTileMatrixSet.isSameProjection();
+            int minZoom = getMinZoom();
             BoundingBox updateBoundingBox = tileBounds.get(minZoom);
             if (!sameProjection) {
                 updateBoundingBox = updateBoundingBox.transform(transformProjectionToTileMatrixSet);