+++ geopackage-sdk/src/main/java/mil/nga/geopackage/tiles/features/FeatureTileGenerator.java
@@ -2,6 +2,8 @@ package mil.nga.geopackage.tiles.features;
 
 import android.content.Context;
 
+import java.util.Collection;
+
 import mil.nga.geopackage.BoundingBox;
 import mil.nga.geopackage.GeoPackage;
 import mil.nga.geopackage.extension.nga.link.FeatureTileTableLinker;
@@ -36,6 +38,43 @@ public class FeatureTileGenerator extends TileGenerator {
      * @param geoPackage   GeoPackage
      * @param tableName    table name
      * @param featureTiles feature tiles
+     * @param boundingBox  tiles bounding box
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, BoundingBox boundingBox,
+                                Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, geoPackage, boundingBox,
+                projection);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevel    zoom level
+     * @param boundingBox  tiles bounding box
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, int zoomLevel, BoundingBox boundingBox,
+                                Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, geoPackage, zoomLevel,
+                boundingBox, projection);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
      * @param minZoom      min zoom
      * @param maxZoom      max zoom
      * @param boundingBox  tiles bounding box
@@ -46,6 +85,87 @@ public class FeatureTileGenerator extends TileGenerator {
                                 String tableName, FeatureTiles featureTiles, int minZoom, int maxZoom, BoundingBox boundingBox, Projection projection) {
         this(context, geoPackage, tableName, featureTiles, geoPackage, minZoom, maxZoom,
                 boundingBox, projection);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevels   zoom levels
+     * @param boundingBox  tiles bounding box
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, Collection<Integer> zoomLevels,
+                                BoundingBox boundingBox, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, geoPackage, zoomLevels,
+                boundingBox, projection);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevels   zoom levels
+     * @param boundingBox  tiles bounding box
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, int[] zoomLevels,
+                                BoundingBox boundingBox, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, geoPackage, zoomLevels,
+                boundingBox, projection);
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param boundingBox       tiles bounding box
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                BoundingBox boundingBox, Projection projection) {
+        super(context, geoPackage, tableName, getBoundingBox(featureGeoPackage,
+                featureTiles, boundingBox, projection), projection);
+        this.featureTiles = featureTiles;
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevel         zoom level
+     * @param boundingBox       tiles bounding box
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                int zoomLevel, BoundingBox boundingBox, Projection projection) {
+        super(context, geoPackage, tableName, zoomLevel,
+                getBoundingBox(featureGeoPackage, featureTiles, boundingBox,
+                        projection),
+                projection);
+        this.featureTiles = featureTiles;
     }
 
     /**
@@ -73,6 +193,85 @@ public class FeatureTileGenerator extends TileGenerator {
     }
 
     /**
+     * Constructor
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevels        zoom levels
+     * @param boundingBox       tiles bounding box
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                Collection<Integer> zoomLevels, BoundingBox boundingBox,
+                                Projection projection) {
+        super(context, geoPackage, tableName, zoomLevels,
+                getBoundingBox(featureGeoPackage, featureTiles, boundingBox,
+                        projection),
+                projection);
+        this.featureTiles = featureTiles;
+    }
+
+    /**
+     * Constructor
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevels        zoom levels
+     * @param boundingBox       tiles bounding box
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                int[] zoomLevels, BoundingBox boundingBox, Projection projection) {
+        super(context, geoPackage, tableName, zoomLevels,
+                getBoundingBox(featureGeoPackage, featureTiles, boundingBox,
+                        projection),
+                projection);
+        this.featureTiles = featureTiles;
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, new int[]{}, null,
+                projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevel    zoom level
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, int zoomLevel, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, zoomLevel, null, projection);
+    }
+
+    /**
      * Constructor, find the the bounding box from the feature table
      *
      * @param context      app context
@@ -94,6 +293,77 @@ public class FeatureTileGenerator extends TileGenerator {
     /**
      * Constructor, find the the bounding box from the feature table
      *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevels   zoom levels
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, Collection<Integer> zoomLevels,
+                                Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, zoomLevels, null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context      app context
+     * @param geoPackage   GeoPackage
+     * @param tableName    table name
+     * @param featureTiles feature tiles
+     * @param zoomLevels   zoom levels
+     * @param projection   tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, int[] zoomLevels,
+                                Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, zoomLevels, null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, featureGeoPackage,
+                new int[]{}, null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevel         zoom level
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                int zoomLevel, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, featureGeoPackage, zoomLevel,
+                null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
      * @param context           app context
      * @param geoPackage        GeoPackage
      * @param tableName         table name
@@ -109,6 +379,44 @@ public class FeatureTileGenerator extends TileGenerator {
                                 int minZoom, int maxZoom, Projection projection) {
         this(context, geoPackage, tableName, featureTiles, featureGeoPackage, minZoom,
                 maxZoom, null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevels        zoom levels
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                Collection<Integer> zoomLevels, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, featureGeoPackage, zoomLevels,
+                null, projection);
+    }
+
+    /**
+     * Constructor, find the the bounding box from the feature table
+     *
+     * @param context           app context
+     * @param geoPackage        GeoPackage
+     * @param tableName         table name
+     * @param featureTiles      feature tiles
+     * @param featureGeoPackage feature GeoPackage if different from the destination
+     * @param zoomLevels        zoom levels
+     * @param projection        tiles projection
+     * @since 6.1.4
+     */
+    public FeatureTileGenerator(Context context, GeoPackage geoPackage, String tableName,
+                                FeatureTiles featureTiles, GeoPackage featureGeoPackage,
+                                int[] zoomLevels, Projection projection) {
+        this(context, geoPackage, tableName, featureTiles, featureGeoPackage, zoomLevels,
+                null, projection);
     }
 
     /**