+++ CHANGELOG.md
@@ -6,7 +6,12 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 6.1.4 (TBD)
 
-* TBD
+* geopackage-core version 6.1.2
+* ormlite-android version 6.1
+* tiff version 3.0.0
+* appcompat version 1.4.1
+* TileGenerator individual zoom level specifications
+* TileCreator dynamic separated tile width and height determination adjustment
 
 ## [6.1.3](https://github.com/ngageoint/geopackage-android/releases/tag/6.1.3) (12-17-2021)